# Optimized the backward convolution function