I've made several optimizations to the backward function in the convolution code; the main areas of improvement include padding, rotation, and dilation operations. These changes aim to reduce computational overhead and a speed up.

**Changes**

_Padding Optimization_
Instead of adding padding to the input data separately, I've integrated the fixed padding values directly into the convolution operation. This reduces the number of steps required.

_Rotation Optimization_
The previous method involved rotating filters by 180 degrees. To optimize this, I replaced the rotation with a combination of matrix flipping or transposing followed by reversing. Additionally I embedded the rotation directly into the convolution step, which simplifies the overall computation.

_Dilation Optimization_
For consistent and predetermined dilation factors, I've embedded these directly into the convolution. This minimizes the total computational workload and enhances performance.

I'm happy to show you these benchmarks run based on the mnist-code from the [mlpack-benchmarks repo](https://github.com/MarkFischinger/mlpack-benchmarks/blob/master/src/convolutional/backward/mnist/mnist_benchmark.cpp):
```
Time Elapsed: 73222ms (original)
Time Elapsed: 50517ms (optimized)
```