@MarkFischinger 2024-06-16
@shrit After a lot testing/debugging, I believe I have finally found the issue. This should resolve it.