@MarkFischinger 2024-06-27
So, I found out that these small adjustments increased the speed by approximately 30.5%.

Here are the MNIST benchmarks:
**Original**
```
756/756 [====================================================================================================] 100% - 59.914s/epoch; 79ms/step; loss: 9.42397
Validation loss: 10191.9.
Time elapsed: 61458ms
Accuracy: train = 70.4048%,      valid = 70.1429%
```


**Optimized**
```
756/756 [====================================================================================================] 100% - 41.1419s/epoch; 54ms/step; loss: 9.42397
Validation loss: 10191.9.
Time elapsed: 42741ms
Accuracy: train = 70.4048%,      valid = 70.1429%
```